+++ Samples~/BasicUsage/AttributeEffect.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Jeomseon.Samples.Attributes
+{
+    [Serializable]
+    public abstract class AttributeEffect
+    {
+        public string Description;
+    }
+}